% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loglinear}
\alias{loglinear}
\title{A multi-purpose log linear model}
\usage{
loglinear(
  formula,
  data,
  latlevels = NULL,
  lambda = 0,
  stderr = c("incomplete", "complete"),
  control = list()
)
}
\arguments{
\item{formula}{formula of the model to be fitted. If the data contain individual records
the lhs of the formula should be left empty.}

\item{data}{a data frame with the data, either with individual records or in the form
of a contingency table. See "Details" for the specification of the variables.}

\item{latlevels}{a vector with the number of levels of the latent variable(s) in
the model. If present, the default \code{NULL} sets the levels of each latent variable to 2.}

\item{lambda}{a scalar denoting the size of the ridge penalty \eqn{\lambda\sum\beta^2}. Defaults to 0.}

\item{stderr}{determines whether the standard error are computed using the incomplete or the
complete data log-likelihood. The latter option is much faster when the model is complex, but
underestimates the standard errors if there are missings and/or latent
variables. If not, both are identical and the complete data log likelihood is used.}

\item{control}{a list with parameters controlling the fitting process:
\itemize{
\item \code{maxit} the maximum number of iterations of the EM algorithm. Defaults to 500.
\item \code{tol} The algorithm converges when \eqn{LL - LL_{old} < tol}. Defaults to 1e-5.
\item \code{seed} seed for obtaining the random starting values. Defaults to a random seed.
\item \code{stepsize} step size for the Newton-Raphson updates of the parameter estimates.
}}
}
\value{
a list with the following components:
\item{formula}{formula used to fit the model.}
\item{coefs}{log linear parameter estimates, se's, and t- and p-values.}
\item{dtable}{data frame in the form of a contingency table used to fit the model.}
\item{dfitted}{data frame the fitted frequencies, including those of the structural zeros.}
\item{D}{the model matrix of the fitted model}
\item{vcv}{variance-covariance matrix of the parameter estimates.}
\item{dimp}{sparse matrix that specifies how the frequencies of the observed (incomplete) data are
to be distributed over the completed data.}
\item{hist}{iteration history.}
}
\description{
\code{loglinear} fits a log linear model with the EM-algoritm. The model
that allows for imputation of missing values, the inclusion of latent variables,
multiple systems estimation and/or regularization.
}
\details{
The function \code{loglinear} interprets the role of the variables in the formula
by their names.

The response variable in the rhs of the formula contains the observed frequencies if the is
in the form of a contingency table, e.g. \code{Freq ~ A + B}. If the data contains individual
records the lhs of the formula  should be left empty, e.g. \code{~ A + B}.

Variables starting with upper case letters A-W are interpreted as population
registers, and these variables should be coded 0 (not observed in the register) and 1 (observed
in the register). If two or more are population registers are present in the model multiple systems
estimation will performed, and a population size estimate will be provided.

Variables named X, Y and/or Z are interpreted as latent variables. If present in the model, the
\code{latlevels} specifies their respective number of levels.

Variables that are neither latent nor population registers should start with a lower case letter.
Variables of class \code{numeric} will be treated as such, but they should have a limited number of
unique values.
}
