% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz13}
\alias{nz13}
\title{New Zealand population data of 2013.}
\format{
Data frame with 923 rows and 9 (categorical) variables.
}
\source{
Stats New Zealand.
}
\usage{
nz13
}
\description{
An extraction of the year 2013 from the data \code{\link{nzy}}. The
variable \code{Y} is not included in the extraction.
}
\details{
The \code{class} of variables with population registers and ethnic
classifications has been changed form \code{numeric} to \code{factor}. See
\code{\link{nzy}} for the variable specifications.
}
\keyword{datasets}
