% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzy}
\alias{nzy}
\title{New Zealand population data of 2013 to 2020.}
\format{
Data frame with 7455 rows and 10 variables.

\describe{
\item{A}{Register A (0 = case not observed in A, 1  = case observed in A)}
\item{B}{Register B (0 = case not observed in B, 1  = case observed in B)}
\item{C}{Register D (0 = case not observed in C, 1  = case observed in C)}
\item{B}{Register D (0 = case not observed in D, 1  = case observed in D)}
\item{a}{ethnic classifications in register A (levels 1, 2, 3, 4)}
\item{b}{ethnic classifications in register B (levels 1, 2, 3, 4)}
\item{c}{ethnic classifications in register C (levels 1, 2, 3, 4)}
\item{d}{ethnic classifications in register D (levels 1, 2, 3, 4)}
\item{y}{the year the observation was made in}
\item{Freq}{The observed frequencies}
}
}
\source{
Stats New Zealand.
}
\usage{
nzy
}
\description{
Contingency table containing population registers and
ethnic classifications over the years 2013 to 2020. The frequencies are
rounded to give effect to the confidentiality provisions of the
Statistics Act of 1975.
}
\details{
The variables \code{a, b, c, d} are partially misclassified and
contain missing values.
}
\keyword{datasets}
