% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzy2}
\alias{nzy2}
\title{Updated version of \code{\link{nzy}} with variables A-D as factors.}
\format{
Data frame with 7028 rows and 10 variables.
}
\source{
Stats New Zealand.
}
\usage{
nzy2
}
\description{
Updated version of \code{\link{nzy}} with variables A-D as factors.
}
\keyword{datasets}
